+++ src/easysam/deploy.py
@@ -15,19 +15,28 @@ import easysam.utils as u
 
 SAM_CLI_VERSION = '1.138.0'
 PIP_VERSION = '25.1.1'
+DEFAULT_SAM_TOOL = 'uv run sam'
 
 
-def deploy(cliparams: dict, directory: Path, deploy_ctx: benedict):
-    '''
+def deploy(
+    toolparams: dict,
+    directory: Path,
+    deploy_ctxs: list[benedict],
+):
+    """
     Deploy a SAM template to AWS.
 
     Args:
-        cliparams: The CLI parameters.
+        toolparams: The CLI parameters.
         directory: The directory containing the SAM template.
-        deploy_ctx: The deployment context.
-    '''
+        deploy_ctxs: The deployment contexts.
 
-    resources, errors = generate(cliparams, directory, [], deploy_ctx)
+    Note: other deployment contexts can be discovered by inspecting the resources.yaml file.
+    """
+
+    resources_set, errors = generate(
+        toolparams, directory, deploy_ctxs
+    )
 
     if errors:
         lg.error(f'There were {len(errors)} errors:')
@@ -35,29 +44,31 @@ def deploy(cliparams: dict, directory: Path, deploy_ctx: benedict):
         for error in errors:
             lg.error(error)
 
-        raise UserWarning('There were errors - aborting deployment')
+        raise UserWarning(
+            f'There were {len(errors)} errors - aborting deployment'
+        )
 
-    lg.info(f'Deploying SAM template from {directory}')
-    check_pip_version(cliparams)
-    check_sam_cli_version(cliparams)
-    remove_common_dependencies(directory)
-    copy_common_dependencies(directory, resources)
+    _check_pip_version(toolparams)
+    _check_sam_cli_version(toolparams)
 
-    # Building the application from the SAM template
-    sam_build(cliparams, directory)
+    for deploy_ctx in deploy_ctxs:
+        name = deploy_ctx.get('name', 'default')
+        lg.info(f'Invoking deployment for {name}')
+        resources = resources_set[name]
 
-    # Deploying the application to AWS
-    sam_deploy(cliparams, directory, deploy_ctx, resources)
+        _deploy_with_context(
+            toolparams=toolparams,
+            resources=resources,
+            directory=directory,
+            deploy_ctx=deploy_ctx,
+        )
 
-    if not cliparams.get('no_cleanup'):
-        remove_common_dependencies(directory)
 
-
-def delete(cliparams, environment):
+def delete(toolparams, environment):
     lg.info(f'Deleting SAM template from {environment}')
-    force = cliparams.get('force')
-    await_deletion = cliparams.get('await_deletion')
-    cf = u.get_aws_client('cloudformation', cliparams)
+    force = toolparams.get('force')
+    await_deletion = toolparams.get('await_deletion')
+    cf = u.get_aws_client('cloudformation', toolparams)
     mode = 'FORCE_DELETE_STACK' if force else 'STANDARD'
     cf.delete_stack(StackName=environment, DeletionMode=mode)  # type: ignore
 
@@ -90,7 +101,28 @@ def delete(cliparams, environment):
     lg.info(f'Stack {environment} deleted')
 
 
-def check_pip_version(cliparams):
+def _deploy_with_context(
+    *,
+    toolparams: dict,
+    resources: benedict,
+    directory: Path,
+    deploy_ctx: benedict,
+):
+    lg.info(f'Deploying SAM template from {directory}')
+    remove_common_dependencies(directory)
+    copy_common_dependencies(directory, resources)
+
+    # Building the application from the SAM template
+    _sam_build(toolparams, directory, deploy_ctx)
+
+    # Deploying the application to AWS
+    _sam_deploy(toolparams, directory, deploy_ctx, resources)
+
+    if not toolparams.get('no_cleanup'):
+        remove_common_dependencies(directory)
+
+
+def _check_pip_version(toolparams):
     lg.info('Checking pip version')
 
     try:
@@ -104,9 +136,9 @@ def check_pip_version(cliparams):
         raise UserWarning('pip not found') from e
 
 
-def check_sam_cli_version(cliparams):
+def _check_sam_cli_version(toolparams):
     lg.info('Checking SAM CLI version')
-    sam_tool = cliparams['sam_tool']
+    sam_tool = toolparams.get('sam_tool', DEFAULT_SAM_TOOL)
     sam_params = sam_tool.split(' ')
     sam_params.append('--version')
 
@@ -122,18 +154,53 @@ def check_sam_cli_version(cliparams):
         raise UserWarning(f'SAM CLI not found. Error: {e}') from e
 
 
-def sam_build(cliparams, directory):
+def _sam_build(
+    toolparams: dict,
+    directory: Path,
+    deploy_ctx: benedict,
+):
     lg.info(f'Building SAM template from {directory}')
-    sam_tool = cliparams['sam_tool']
-    sam_params = sam_tool.split(' ')
-    sam_params.append('build')
+    build_dir = u.get_build_dir(directory, deploy_ctx)
+    template = Path(build_dir, 'template.yml')
 
-    if cliparams.get('verbose'):
+    if not template.exists():
+        raise UserWarning(
+            f'Template {template} not found in build directory {build_dir}'
+        )
+
+    sam_tool: str = toolparams.get('sam_tool', DEFAULT_SAM_TOOL)
+    sam_params = sam_tool.split(' ')
+
+    sam_build_dir = Path(
+        build_dir.relative_to(directory), '.aws-sam'
+    )
+
+    sam_params.extend(
+        [
+            'build',
+            '--template-file',
+            template.relative_to(directory).as_posix(),
+            '--build-dir',
+            sam_build_dir.as_posix(),
+        ]
+    )
+
+    if region := deploy_ctx.get('target_region'):
+        sam_params.extend(['--region', region])
+    else:
+        lg.info('No AWS region found in the target. Relying on the environment or profile to infer the region.')
+
+    if toolparams.get('verbose'):
         sam_params.append('--debug')
 
     try:
-        lg.debug(f'Running command: {" ".join(sam_params)}')
-        subprocess.run(sam_params, cwd=directory.resolve(), text=True, check=True)
+        lg.info(f'Running command: {" ".join(sam_params)}')
+        subprocess.run(
+            sam_params,
+            cwd=directory.as_posix(),
+            text=True,
+            check=True,
+        )
         lg.info('Successfully built SAM template')
 
     except subprocess.CalledProcessError as e:
@@ -141,57 +208,79 @@ def sam_build(cliparams, directory):
         raise UserWarning('Failed to build SAM template') from e
 
 
-def sam_deploy(cliparams, directory, deploy_ctx, resources):
-    lg.info(f'Deploying SAM template from {directory} to\n{json.dumps(deploy_ctx, indent=4)}')
-    sam_tool = cliparams['sam_tool']
+def _sam_deploy(toolparams, directory, deploy_ctx, resources):
+    lg.info(
+        f'Deploying SAM template from {directory} to\n{json.dumps(deploy_ctx, indent=4)}'
+    )
+    sam_tool = toolparams.get('sam_tool', DEFAULT_SAM_TOOL)
     sam_params = sam_tool.split(' ')
-
     aws_stack = deploy_ctx['environment']
 
     if not aws_stack:
         raise UserWarning('No AWS stack found in deploy context')
 
-    sam_params.extend([
-        'deploy',
-        '--parameter-overrides', f'ParameterKey=Stage,ParameterValue={aws_stack}',
-        '--stack-name', aws_stack,
-        '--no-fail-on-empty-changeset',
-        '--no-confirm-changeset',
-        '--resolve-s3',
-        '--capabilities', 'CAPABILITY_IAM', 'CAPABILITY_NAMED_IAM'
-    ])
+    sam_params.extend(
+        [
+            'deploy',
+            '--parameter-overrides',
+            f'ParameterKey=Stage,ParameterValue={aws_stack}',
+            '--stack-name',
+            aws_stack,
+            '--no-fail-on-empty-changeset',
+            '--no-confirm-changeset',
+            '--resolve-s3',
+            '--capabilities',
+            'CAPABILITY_IAM',
+            'CAPABILITY_NAMED_IAM',
+        ]
+    )
 
     region = deploy_ctx.get('target_region')
 
     if region:
         sam_params.extend(['--region', region])
+    else:
+        lg.info('No AWS region found in the target. Relying on the environment or profile to infer the region.')
 
-    aws_tags = list(cliparams.get('tag', []))
+    aws_tags = list(toolparams.get('tag', []))
 
     if 'tags' in resources:
-        aws_tags.extend(
-            f'{name}={value}'
-            for name, value in resources['tags'].items()
-        )
+        aws_tags.extend(f'{name}={value}' for name, value in resources['tags'].items())
 
     aws_tag_string = ' '.join(aws_tags)
     lg.info(f'AWS tag string: {aws_tag_string}')
     sam_params.extend(['--tags', aws_tag_string])
 
-    if cliparams.get('verbose'):
+    if toolparams.get('verbose'):
         sam_params.append('--debug')
 
-    if cliparams['dry_run']:
-        lg.info(f'Would run: {" ".join(sam_params)}')
+    if target_profile := deploy_ctx.get('target_profile'):
+        lg.info(
+            f'Using AWS profile from target: {target_profile}'
+        )
+        sam_params.extend(['--profile', target_profile])
+    else:
+        lg.warning('No AWS profile found in target. Relying on the environment to infer the profile.')
+
+    build_dir = u.get_build_dir(directory, deploy_ctx)
+    sam_build_dir = Path(build_dir, '.aws-sam')
+    run_description = f'{" ".join(sam_params)} in {sam_build_dir}'
+
+    if 'dry_run' not in toolparams:
+        raise UserWarning('dry_run must be present in `toolparams`')
+
+    if toolparams['dry_run']:
+        lg.info(f'Would run: {run_description}')
         return
 
-    if cliparams.get('aws_profile'):
-        lg.info(f'Using AWS profile: {cliparams["aws_profile"]}')
-        sam_params.extend(['--profile', cliparams['aws_profile']])
-
     try:
-        lg.debug(f'Running command: {" ".join(sam_params)}')
-        subprocess.run(sam_params, cwd=directory.resolve(), text=True, check=True)
+        lg.info(f'Running command: {run_description}')
+        subprocess.run(
+            sam_params,
+            cwd=sam_build_dir.as_posix(),
+            text=True,
+            check=True,
+        )
         lg.info('Successfully deployed SAM template')
 
     except subprocess.CalledProcessError as e: